refactor(shim-client,studio,workspace): shared shim export
to avoid extra imports into that package accidentally
