+++ packages/shared/src/types/index.ts
@@ -1,3 +1,2 @@
-export type * from "./shim";
 export type * from "./telemetry";
 export type * from "./workspace";
